requirements.txt
Flask==2.0.1
Jinja2==3.0.1
python-dotenv==0.18.0
requests==2.25.1
